hyperFakeModel/pyproject.toml
[project]
name = "hyperfakeworker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.2.1",
    "joblib>=1.5.1",
    "jupyterlab>=4.4.2",
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "pandas>=2.2.3",
    "polars>=1.30.0",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "torch>=2.7.0",
    "tqdm>=4.67.1",
]
